finish api catalog

+++ src/main/java/n1h5/models/service/AuditLogService.java
@@ -13,7 +13,7 @@ public class AuditLogService {
     private final AuditLogRepository auditLogRepository;
 
     @Async // Chạy bất đồng bộ để không ảnh hưởng tới tốc độ phản hồi API
-    public void saveLog(Long userId, String action, String entityName, Long entityId, String oldValue, String newValue) {
+    public void saveLog(Integer userId, String action, String entityName, Long entityId, String oldValue, String newValue) {
         AuditLog log = AuditLog.builder()
                 .userId(userId)
                 .action(action)
@@ -22,7 +22,6 @@ public class AuditLogService {
                 .oldValue(oldValue)
                 .newValue(newValue)
                 .build();
-
         auditLogRepository.save(log);
     }
 }
